Solved issue with profile savings, moved it to appdata. New installer with .exe and .msi available!

+++ Midi2Vol/Config.cs
@@ -1,26 +1,40 @@
 ﻿using Newtonsoft.Json;
 using System;
 using System.Collections.Generic;
-
+using System.IO;
 
 namespace Midi2Vol
 {
     class Config
     {
-        public String configFile = "config.json";
+        public String configFile = Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData)+"/Midi2Vol/config.json";
         public Config() { }
         public List<App> SourceConfig()
         {
-            String jsonString = System.IO.File.ReadAllText(configFile);
-            List<App> apps = JsonConvert.DeserializeObject<List<App>>(jsonString);
-            return apps;
+            FileAttributes fileAttributes = File.GetAttributes(configFile);
+            if ((fileAttributes & FileAttributes.ReadOnly) != FileAttributes.ReadOnly)
+            {
+                String jsonString = System.IO.File.ReadAllText(configFile);
+                List<App> apps = JsonConvert.DeserializeObject<List<App>>(jsonString);
+                return apps;
+            }
+            return null;
         }
         public void saveConfig(List<App> apps)
         {
             String objectJson = JsonConvert.SerializeObject(apps);
-            System.IO.File.WriteAllText(configFile, objectJson);
-
-
+            FileAttributes fileAttributes = File.GetAttributes(configFile);
+            if ((fileAttributes & FileAttributes.ReadOnly) != FileAttributes.ReadOnly)
+            {
+                try
+                {
+                    System.IO.File.WriteAllText(configFile, objectJson);
+                }
+                catch
+                {
+                    Console.WriteLine("Could not get Stream for '" + configFile + "'. Possibly in use");
+                }
+            }
         }
     }
 }
